Rename tensor4all-tensortrain to tensor4all-simpletensortrain and integrate MPO contraction
## Changes

- Rename crate `tensor4all-tensortrain` to `tensor4all-simpletensortrain`
- Integrate `tensor4all-mpocontraction` as `mpo` submodule in simpletensortrain
- Remove `tensor4all-mpocontraction` crate (merged into simpletensortrain)
- Remove hadamard-related functionality from contraction.rs (keep only dot product)
- Disable C-API, Julia, and Python bindings for tensortrain/MPO (pending re-integration)

## Details

### Crate Rename
- Directory: `tensor4all-tensortrain/` → `tensor4all-simpletensortrain/`
- All internal references updated to new crate name

### MPO Integration
- Moved mpocontraction source files to `src/mpo/` submodule
- Fixed import paths (`crate::` → `super::` for submodule context)
- Re-exports available via `tensor4all_simpletensortrain::mpo`
- Deleted original mpocontraction crate

### Disabled Bindings (files deleted, code preserved in git history)
- C-API: tensortrain.rs, mpocontraction.rs
- Julia: TensorTrain/, MPO/
- Python: tensortrain/, mpo/
