+++ Cargo.toml
@@ -8,9 +8,9 @@ members = [
     "crates/tensor4all-treetn",
     "crates/quanticsgrids",
     "crates/tensor4all-matrixci",
-    "crates/tensor4all-tensortrain",
+    "crates/tensor4all-simpletensortrain",
     "crates/tensor4all-tensorci",
-    "crates/tensor4all-mpocontraction",
+    # "crates/tensor4all-mpocontraction",  # Merged into tensor4all-simpletensortrain
     "crates/tensor4all-capi",
 ]
 resolver = "2"
