+++ crates/tensor4all-capi/Cargo.toml
@@ -15,8 +15,9 @@ crate-type = ["cdylib", "rlib"]
 tensor4all-core-common = { path = "../tensor4all/core-common" }
 tensor4all-core-tensor = { path = "../tensor4all/core-tensor" }
 tensor4all-core-linalg = { path = "../tensor4all/core-linalg" }
-tensor4all-tensortrain = { path = "../tensor4all-tensortrain" }
-tensor4all-mpocontraction = { path = "../tensor4all-mpocontraction" }
+# Disabled - pending simpletensortrain integration
+# tensor4all-simpletensortrain = { path = "../tensor4all-simpletensortrain" }
+# tensor4all-mpocontraction = { path = "../tensor4all-mpocontraction" }
 libc = "0.2"
 thiserror.workspace = true
 paste = "1.0"
