+++ crates/tensor4all-simpletensortrain/src/arithmetic.rs
