+++ crates/tensor4all-simpletensortrain/src/cache.rs
