+++ crates/tensor4all-simpletensortrain/src/canonical.rs
