+++ crates/tensor4all-simpletensortrain/src/compression.rs
