+++ crates/tensor4all-simpletensortrain/src/error.rs
