+++ crates/tensor4all-simpletensortrain/src/mpo/error.rs
