+++ crates/tensor4all-simpletensortrain/src/traits.rs
