+++ crates/tensor4all-simpletensortrain/src/types.rs
