+++ crates/tensor4all-simpletensortrain/src/vidal.rs
