--- python/tensor4all/src/tensor4all/mpo/__init__.py
@@ -1,8 +1,0 @@
-"""MPO (Matrix Product Operator) module.
-
-This module provides MPO types and operations for tensor network computations.
-"""
-
-from ._mpo import MPOF64, MPOC64, ContractionAlgorithm
-
-__all__ = ["MPOF64", "MPOC64", "ContractionAlgorithm"]
